+++ src/main/kotlin/com/kotcrab/ghidra/rest/model/ApiBookmark.kt
@@ -5,6 +5,7 @@ import java.math.BigInteger
 data class ApiBookmark(
   val id: Long,
   val address: BigInteger,
+  val addressSpaceId: Int,
   val type: String,
   val category: String,
   val comment: String,