+++ src/main/kotlin/com/kotcrab/ghidra/rest/model/ApiRelocation.kt
@@ -5,6 +5,7 @@ import java.math.BigInteger
 
 data class ApiRelocation(
   val address: BigInteger,
+  val addressSpaceId: Int,
   val status: Relocation.Status,
   val type: Int,
   val values: LongArray,